services/python/requirements.txt
pygetwindow==0.0.9
Pillow>=10.3.0
numpy>=1.21.0,<2.0.0
tensorflow>=2.12.0,<2.16.0
matplotlib>=3.5.0
pandas>=1.5.0
opencv-python>=4.8.0
